requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
aiofiles
pydantic[email]
passlib
bcrypt
python-jose

